REFACTOR RUSTAUTFILL
